Move StepRoute component into its own file
