+++ client/landing/stepper/utils/kebabCase.ts
@@ -1,0 +1,3 @@
+const kebabCase = ( value: string ) => value.replace( /([a-z0-9])([A-Z])/g, '$1-$2' ).toLowerCase();
+
+export default kebabCase;
